backend/requirements.txt
fastapi>=0.100.0
uvicorn==0.22.0
sqlalchemy>=2.0.27
pydantic>=2.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
psycopg2-binary==2.9.6 # PostgreSQL用ドライバー
pymongo==4.3.3
